fix(a3p): remove usage of duplicated code
